Support for user-friendly github releases
In the flavour of the rest of the package.  It will look in the DESCRIPTION
file for the current version and make a release based on that.  Checks are
done to make sure that the upstream code is up to date, that the local repo
is clean and that the version number has been incrememented.